[Chore #163208486] Refactor code
