+++ server/test/unit/middlewares/schema/schema.spec.js
@@ -1,0 +1,9 @@
+import 'chai/register-should';
+import schemaValidator from '../../../../middlewares/schema/schemaValidator';
+
+describe.only('Schema Middleware', () => {
+  it('should return a middleware', () => {
+    const validateMiddleware = schemaValidator(true);
+    validateMiddleware.should.be.a('function');
+  });
+});
